requirements.txt
PyQt6>=6.6.0
qasync>=0.27.1
playwright>=1.40.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
aiohttp>=3.9.0
Pillow>=10.0.0
